% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_Oliveira_Ceretta_S_Out_Inv.R
\name{ANNt_Oliveira_Ceretta_S_Out_Inv}
\alias{ANNt_Oliveira_Ceretta_S_Out_Inv}
\title{ANNt_Oliveira_Ceretta_Out_Inv
Create a portfolio to outperform the Sharpe portfolio of reported assets out of sample.}
\usage{
ANNt_Oliveira_Ceretta_S_Out_Inv(
  Tickers,
  RM,
  Rf,
  Initial_Date,
  Final_Date_Training,
  Final_Date,
  Periodicity,
  Hidden,
  Stepmax,
  Loss = "MSE",
  Learning_Rate = 0.3,
  Decay = "No",
  Early_Stopping = "No",
  Asymmetry = "Negative",
  Type_ANNt = "T4",
  N_Assets,
  Base = "yahoo",
  Import = "Yes",
  Exclude_ticket = "",
  AQ = "A",
  Size = 2000,
  PE_Ratio = 15,
  PB_Ratio = 1.5,
  GI_min = 0,
  GI_max = 21.5,
  CR = 2,
  EPS = 0,
  Plot_IS = "Total Revenue",
  Plot_CF = "Cash Dividends Paid",
  Plot_BS = "Total Liabilities",
  Type_ANN = "ANNt",
  Order = "Yes",
  Skew_t = "No"
)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets}

\item{RM}{Proxy of the market}

\item{Rf}{Risk free rate}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day'). Assets with values not observed in the series are excluded}

\item{Final_Date}{End date of the treatment series}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”}

\item{Hidden}{Number of hidden neurons (If ” is the length series). For a good performance use '' to form a square input x hidden matrix of neurons}

\item{Stepmax}{Number of replications per asset to train the ANN. For a good performance, use 7500}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{Type_ANNt}{Select type ANNt: "T1"= NNet_Signal_Traning; "T2"= NNet_t_Training; "T3"= MC_Signal_Training; "T4"= MC_t_Training; "T5"= NNet_Signal_Test; "T6"= NNet_t_Test; "T7"= MC_Signal_Test; "T8"= Type_ANNt: MC_t_Test}

\item{N_Assets}{Limit of asset numbers in the portfolio}

\item{Base}{Database to use: "yahoo" or "Rus"}

\item{Import}{Import dates from external data base after first import. "Yes"
or "No". "Yes" is the standard.}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}

\item{AQ}{Annual Statements (A) or Quarterly Statements (Q)}

\item{Size}{Minimum Total Assets in Millions}

\item{PE_Ratio}{Maximum Price/Earnings Ratio}

\item{PB_Ratio}{Maximum Price to Book Ratio}

\item{GI_min}{Graham indicator minimum}

\item{GI_max}{Graham indicator maximum}

\item{CR}{Minimum Current Ratio}

\item{EPS}{Earning per Share Minimum}

\item{Plot_IS}{Chart of Income Sheet account selected}

\item{Plot_CF}{Chart of Cash Flow account selected}

\item{Plot_BS}{Char of Balance Sheet accoun selected}

\item{Type_ANN}{Select the network type: 'ANNt', 'LSTMt' in RNN from ANNt, or 'SKEWt' to raw excess return data}

\item{Order}{If "Yes" processes the asset selection, if "No" uses the already processed assets available in the database}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}

\item{Initial_Date_Training}{Training series start Date}

\item{Asymmetric}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative".}
}
\description{
Command that realize all operations of the package and save all in a specific past into user past
}
\examples{
Tickers <-c('AAPL','XOM','TSLA','KO', 'F')
RM <-c('^GSPC') #RM the S&P500
Rf <- 0
Initial_Date <-c('2018-01-03')
Final_Date_Training <- c('2022-12-29')
Final_Date <-c('')
Periodicity <- c('daily')
Hidden <- 5
Stepmax <- 7500
Type_ANNt <- 'T8'
N_Assets <- 3
AQ='A'
Size=2000
PE_Ratio=15
PB_Ratio=1.5
GI_min=0
GI_max=21.5
CR=2
EPS=0
Plot_IS=''
Plot_BS = ''
Plot_CF=''
ANNt_Oliveira_Ceretta_S_Out_Inv(c('AAPL','XOM','TSLA','KO', 'F'), '^GSPC', 0,
'2018-01-03', '2022-12-29', '', 'daily',5,7500,'T8',3,
AQ='A', Size=2000, PE_Ratio=315, PB_Ratio=211.5,GI_min=0, GI_max=2100.5,
 CR=1, EPS=0, Plot_IS='', Plot_CF='', Plot_BS='')
}
