% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backup_Inv.R
\name{Backup_Inv}
\alias{Backup_Inv}
\title{Backup_Inv
Generate backup of the role}
\usage{
Backup_Inv(Investment = "No")
}
\arguments{
\item{Investment}{'No' is default; 'Yes' if command Investment_Horizon was executed}

\item{()}{No require parameters}
}
\description{
Backup_Inv
Generate backup of the role
}
\examples{
Backup_Inv()
}
