% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exclude_Tickers.R
\name{Exclude_Tickers}
\alias{Exclude_Tickers}
\title{Exclude_Tickers
#' Exclud tickers that not have cash flow statment published in FinVis}
\usage{
Exclude_Tickers(Tickers, AQ = "A", FS = "C", Exclude_ticket = "")
}
\arguments{
\item{Tickers}{to import and analyzes the financial parameters.
For all SP500 use 'Current_SP500_Tickers'}

\item{AQ}{Annual Statements (A) or Quarterly Statements (Q)}

\item{FS}{= Financial Statement (FS): Income Statement (I), Balance Sheet (B), Cash Flow (C)}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}
}
\description{
Exclude_Tickers
#' Exclud tickers that not have cash flow statment published in FinVis
}
\examples{
Exclude_Tickers('Current_SP500_Tickers', 'Q', 'C')
}
