% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_Portfolios_Inv.R
\name{Gen_Portfolios_Inv}
\alias{Gen_Portfolios_Inv}
\title{Gen_Portfolios_Inv Generate investment portfolios specified
Generate portfolios selected, how Sharpe, Markowitz, ANNt, Magic Formula, Intelligent Investor}
\usage{
Gen_Portfolios_Inv(
  Portfolios,
  N_Assets,
  Initial_Date_Testing,
  Final_Date_Testing,
  Rf,
  type_ANNt
)
}
\arguments{
\item{Portfolios}{Specified portfolios generated: "Magic_Formula" or "Intelligent_Investor"}

\item{N_Assets}{limit of asset numbers in the portfolio}

\item{Initial_Date_Testing}{Initial Date of Test Series}

\item{Final_Date_Testing}{Final Date Test Series, if '' is the system date}

\item{Rf}{Risk free rate}

\item{type_ANNt}{Select type ANNt:
"T1"= NNet_Signal_Traning;
"T2"= NNet_t_Training;
"T3"= MC_Signal_Training;
"T4"= MC_t_Training;
"T5"= NNet_Signal_Test;
"T6"= NNet_t_Test;
"T7"= MC_Signal_Test;
"T8"= Type_ANNt: MC_t_Test}
}
\description{
Gen_Portfolios_Inv Generate investment portfolios specified
Generate portfolios selected, how Sharpe, Markowitz, ANNt, Magic Formula, Intelligent Investor
}
\examples{
Portfolios <- c('Magic_Formula', 'Intelligent_Investor')
N_Assets <- 3
Initial_Date_Testing <- c('2023-01-03')
Final_Date_Testing <- c('')
Rf <- 0
type_ANNt <- 'T8'
# Generate assets portfolio (maximum N assets specified)
Gen_Portfolios_Inv(c('Magic_Formula', 'Intelligent_Investor'),20,'2023-01-03','',0,'T8')
}
\author{
Alexandre Silva de Oliveira
}
