% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFins.R
\name{getFins}
\alias{getFins}
\title{getFins - Get Financials
Income Statement, Cash Flow Statment, & Balance Sheet}
\usage{
getFins(symbol, AQ, FS)
}
\arguments{
\item{symbol}{= ticker symbol}

\item{AQ}{= Annual (A) or Quarterly (Q)?}

\item{FS}{= Financial Statement (FS): Income Statement (I), Balance Sheet (B), Cash Flow (C)}
}
\description{
getFins - Get Financials
Income Statement, Cash Flow Statment, & Balance Sheet
}
\examples{
getFins('TSLA', 'A', 'B')
}
