% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graham_Portfolio.R
\name{Graham_Portfolio}
\alias{Graham_Portfolio}
\title{Graham_Portfolio
Return Graham portfolio from Yahoo data
Attend the restriction give for Graham in the Intelligent Investor}
\usage{
Graham_Portfolio(Symbols)
}
\arguments{
\item{Symbols}{Tickers to import and analyzes the financial parameters.
For all SP500 use 'Current_SP500_Tickers'}
}
\description{
Graham_Portfolio
Return Graham portfolio from Yahoo data
Attend the restriction give for Graham in the Intelligent Investor
}
\examples{
# example code
Graham_Portfolio(c("XOM","MSFT","JNJ","GE","CVX","WFC","PG","JPM","VZ","PFE","T","IBM","MRK","BAC","DIS","ORCL","PM","INTC","SLB"))
}
