% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graham_Portfolio_Vis.R
\name{Graham_Portfolio_Vis}
\alias{Graham_Portfolio_Vis}
\title{Graham_Portfolio_Vis
Return Graham portfolio from FinVis data
Attend the restriction give for Graham in the Intelligent Investor}
\usage{
Graham_Portfolio_Vis(
  Tickers,
  AQ = "A",
  Size = 2000,
  PE_Ratio = 15,
  PB_Ratio = 1.5,
  GI_min = 0,
  GI_max = 21.5,
  CR = 2,
  EPS = 0,
  Break = 15,
  Plot_IS = "Total Revenue",
  Plot_CF = "Cash Dividends Paid",
  Plot_BS = "Total Liabilities",
  Exclude_ticket = ""
)
}
\arguments{
\item{Ticker}{to import and analyzes the financial parameters.
For all SP500 use 'Current_SP500_Tickers'}
}
\description{
Graham_Portfolio_Vis
Return Graham portfolio from FinVis data
Attend the restriction give for Graham in the Intelligent Investor
}
