% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_Bovespa_Brazil.R
\name{Index_Bovespa_Brazil}
\alias{Index_Bovespa_Brazil}
\title{Index_Bovespa_Brazil}
\usage{
Index_Bovespa_Brazil()
}
\arguments{
\item{()}{Without parameters}
}
\description{
Import Ibovespa Composition from B3 and investing.com in the last stock exchange floor.
}
\examples{
Index_Bovespa_Brazil()
}
