% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Investment_Horizon_Int.R
\name{Investment_Horizon_Int}
\alias{Investment_Horizon_Int}
\title{Investment_Horizon_Int}
\usage{
Investment_Horizon_Int(
  Tickers,
  RM,
  Rf,
  Initial_Date,
  Final_Date_Training,
  Final_Date,
  Frequency,
  Periodicity,
  Hidden,
  Stepmax,
  Loss = "MSE",
  Learning_Rate = 0.3,
  Decay = "No",
  Early_Stopping = "No",
  Asymmetry = "Negative",
  Type_ANNt = "T4",
  N_Assets,
  Base = "yahoo",
  Fun = "S_Out",
  Skew_t = "No",
  Specific_Dates = Sys.Date(),
  Import = "Yes",
  Exclude_ticket = "",
  AQ = "A",
  Size = 2000,
  PE_Ratio = 15,
  PB_Ratio = 1.5,
  GI_min = 0,
  GI_max = 21.5,
  CR = 2,
  EPS = 0,
  Plot_IS = "Total Revenue",
  Plot_CF = "Cash Dividends Paid",
  Plot_BS = "Total Liabilities",
  Download = "Yes",
  Type_ANN = "ANNt",
  Order = "Yes",
  Continue_from = "1900-01-01"
)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets.}

\item{RM}{Proxy of the market.}

\item{Rf}{Risk free rate.}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day'). Assets with values not observed in the series are excluded.}

\item{Final_Date}{End date of the treatment series.}

\item{Frequency}{How many times simulate the investiment horizon.}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”.}

\item{Hidden}{Number of hidden neurons (If ” is the length series). For a good performance use '' to form a square input x hidden matrix of neurons.}

\item{Stepmax}{Number of replications per asset to train the ANN. For a good performance, use 7500.}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{N_Assets}{Limit of asset numbers in the portfolio.}

\item{Base}{Database to use: "yahoo" or "Rus".}

\item{Fun}{Which technique to apply to generate portfolios:
'S_Out' uses the ANNt_Oliveira_Ceretta_S_Out function, this is the standard;
'Out' uses the ANNt_Oliveira_Cereta_Out function;
'S' uses the ANNt_Oliveira_Cereta_S function;
'Original' uses the ANNt_Oliveira_Ceretta function.}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}

\item{Import}{Import dates from external data base after first import in each re balance. "Yes"
or "No". "No" is the standard.}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}

\item{AQ}{Annual Statements (A) or Quarterly Statements (Q)}

\item{Size}{Minimum Total Assets in Millions}

\item{PE_Ratio}{Maximum Price/Earnings Ratio}

\item{PB_Ratio}{Maximum Price to Book Ratio}

\item{GI_min}{Graham indicator minimum}

\item{GI_max}{Graham indicator maximum}

\item{CR}{Minimum Current Ratio}

\item{EPS}{Earning per Share Minimum}

\item{Plot_IS}{Chart of Income Sheet account selected}

\item{Plot_CF}{Chart of Cash Flow account selected}

\item{Plot_BS}{Char of Balance Sheet account selected}

\item{Download}{Download asset prices for external data set or local data set. "Yes" or "No". "Yes" is the standard.}

\item{Type_ANN}{Select the network type: 'ANNt' or 'LSTMt' in RNN from ANNt}

\item{Order}{If "Yes" processes the asset selection, if "No" uses the already processed assets available in the database}

\item{Continue_from}{Determine if continue from a Specific_Date in the data}

\item{Initial_Date_Training}{Training series start Date.}

\item{Asymmetric}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative".}

\item{Specific_Date}{Specific dates for the end of training. Used to define
the investment horizon of portfolios from specific dates.}
}
\description{
Analyzes the return of portfolios with different investment horizons
}
\examples{
Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
Specific_Dates=c(
'2025-09-30','2025-06-30',
'2025-03-31','2024-12-31',
'2024-09-30','2024-06-30',
'2024-03-31','2023-12-31',
'2023-09-30','2023-06-30',
'2023-03-31','2022-12-31'
)
#######
Investment_Horizon_Int(
Tickers = c('Current_SP500_Tickers'),
RM = c('^GSPC'),
Rf = 0,
Initial_Date = c('2018-01-03'),
Final_Date_Training = c('2019-12-31'),
Final_Date = '',
Frequency = 2,
Periodicity = c('daily'),
Hidden = 7,
Stepmax = 300,
Loss="GMADL",
Learning_Rate=0.3,
Decay=c('Yes',0.1),
Early_Stopping = c('Yes',0.002),
Asymmetry='Positive',
Type_ANNt = 'T2',
N_Assets = 20,
Base = 'yahoo',
Fun = 'S_Out',
Specific_Dates = Specific_Dates,
AQ ='Q',
Size = 2000,
PE_Ratio = 150,
PB_Ratio = 11.5,
GI_min = 0,
GI_max = 2100.5,
CR = 0,
EPS = 0,
Plot_IS = '',
Plot_BS = '',
Type_ANN = 'ANNt',
Plot_CF = '',
Import = 'No',
Exclude_ticket = c('Q','BIIB', 'DVA', 'EPAM', 'DLTR', 'CLX', 'FICO', 'DG','FIS','HIG'),
Download = 'No',
Skew_t = 'Yes')
}
