% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Magic_Formula.R
\name{Magic_Formula}
\alias{Magic_Formula}
\title{Magic_Formula Generate Magic Formula Portfolio
The Joel Greenblatt´s Magic Formula for beating the stock market}
\usage{
Magic_Formula(
  Tickers,
  AQ = "A",
  Break = 20,
  Plot_IS = "Total Revenue",
  Plot_CF = "Cash Dividends Paid",
  Plot_BS = "Total Liabilities",
  Exclude_ticket = ""
)
}
\arguments{
\item{Tickers}{to import and analyzes the financial parameters.
For all SP500 use 'Current_SP500_Tickers'}

\item{AQ}{Annual Statements (A) or Quarterly Statements (Q)}

\item{Break}{Portfolio asset numbers to be generated}

\item{Plot_IS}{Chart of Income Sheet account selected}

\item{Plot_CF}{Chart of Cash Flow account selected}

\item{Plot_BS}{Char of Balance Sheet account selected}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}
}
\description{
Magic_Formula Generate Magic Formula Portfolio
The Joel Greenblatt´s Magic Formula for beating the stock market
}
\examples{
Tickers = 'Current_SP500_Tickers'
AQ = 'A'
GI_min = 0
GI_max = 21.5
EPS = 0
Exclude_ticket = c('AZO','GL','TPL')
Magic_Formula(Tickers='Current_SP500_Tickers',AQ='A',Break=20, Plot_IS='', Plot_BS='',Plot_CF='',Exclude_ticket = c('AZO','GL','TPL'))

}
