% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctBS.R
\name{pctBS}
\alias{pctBS}
\title{pctBS - Porcentage Balance Sheet
Convert Financial Tables to Percentages}
\usage{
pctBS(BS)
}
\arguments{
\item{BS}{= balance sheet from FinViz}
}
\description{
pctBS - Porcentage Balance Sheet
Convert Financial Tables to Percentages
}
\examples{
pctBS('BS')
}
