% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Ratio_Horizon_Inv.R
\name{Plot_Ratio_Horizon_Inv}
\alias{Plot_Ratio_Horizon_Inv}
\title{Plot_Ratio_Horizon_Inv}
\usage{
Plot_Ratio_Horizon_Inv(
  Ratio = "Annualized_Returns",
  Legend_position = "topleft",
  Compare = "RM",
  Until_Date = ""
)
}
\arguments{
\item{Compare}{with:
RM - Return of Market proxy;
MARKOWITZ;
SHARPE;
MF_EQ;
MF_MKW;
MF_SHARPE;
ANNt-EQ;
ANNt_MKW;
ANNt_SHARPE;
Magic_EQ;
Magic_MKW;
Magic_SHARPE;
Graham_EQ;
Graham_MKW;
Graham_SHARPE.}

\item{Until_Date}{Period´s length to plot.}

\item{Ratio:}{Rm - Average_Return - Means of Returns from each portfolio;
Annualized_Returns - This is the standard ratio;
RCum - Cumulative_Returns;
Annualized_Volatility - Annualized Volatility;
Var - Variance with 95% confidence;
CVar - Conditional Variance 95% confidence;
Sharpe - Sharpe Ratio;
Alpha - Jensen's Alpha;
Beta - Covariance between portfolio and market proxy over market proxy Variance;
Sortino - Sortino Ratio;
Treynor - Treynor Ratio.}

\item{Legend_position:}{topleft;
topright;
bottomleft;
bottomrigh;
bottom;
top;
left;
right.}
}
\description{
Analyzes ratios chart of portfolios with different investment horizons.
You need to run the Investment_Horizon command first.
}
\examples{
Plot_Ratio_Horizon_Inv(Ratio="Annualized_Returns")
}
