% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_Backtesting_Inv.R
\name{Portfolio_Backtesting_Inv}
\alias{Portfolio_Backtesting_Inv}
\title{Portfolio_Backtesting_Inv
Realize the portfolio backtest.}
\usage{
Portfolio_Backtesting_Inv(Date_Initial_Backtesting, Date_Final_Backtesting)
}
\arguments{
\item{Date_Initial_Backtesting}{Date initial of the backtest}

\item{Date_Final_Backtesting}{Date final of the backtest}
}
\description{
Portfolio_Backtesting_Inv
Realize the portfolio backtest.
}
\examples{
Date_Initial_Backtesting =''
Date_Final_Backtesting =''
Portfolio_Backtesting_Inv('', '')

}
