% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2plot.R
\name{table2plot}
\alias{table2plot}
\title{table2plot
Plot Quarterly Values from the Financial}
\usage{
table2plot(ticker, WHAT, FROM)
}
\arguments{
\item{ticker}{= ticker symbol}

\item{WHAT}{= One of the row names from the financial table ex. 'Cash & Short Term Investments'}

\item{FROM}{= data generate with getFins Command}
}
\description{
table2plot
Plot Quarterly Values from the Financial
}
\examples{
table2plot('TSLA','Cash & Short Term Investments',bs)
}
