% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turnover.R
\name{Turnover}
\alias{Turnover}
\title{Turnover
It measures portfolio turnover}
\usage{
Turnover(Portfolio)
}
\arguments{
\item{Portfolios:Specify}{the portfolios or type 'Alls' for all.
Portfolios:'MF_EQ', 'MF_MKW','MKW','ANNt_EQ','ANNt_MKW','Sharpe','MF_Sharpe','ANNt_Sharpe',
Magic_EQ','Magic_MKW', 'Magic_Sharpe', 'Graham_EQ', 'Graham_MKW', 'Graham_Sharpe'}
}
\description{
Turnover
It measures portfolio turnover
}
\examples{
Turnover('Alls')
}
