% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights_Investment_Horizon_Inv.R
\name{Weights_Investment_Horizon_Inv}
\alias{Weights_Investment_Horizon_Inv}
\title{Weights_Investment_Horizon_Inv}
\usage{
Weights_Investment_Horizon_Inv(Portfolio = "Sharpe")
}
\arguments{
\item{Portfolio:}{MF_EQ; MF_MKW; MKW; ANNt_EQ; ANNt_MKW; Sharpe; MF_Sharpe; ANNt_Sharpe
Magic_EQ; Magic_MKW; Magic_SHARPE; Graham_EQ; Graham_MKW; or Graham_SHARPE}
}
\description{
Return with weights of the portfolio selected. It is necessary to run the Investiment_Horizon command before.
}
\examples{
Weights_Investment_Horizon_Inv(Portfolio='Sharpe')
}
