% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intelligent_Investor.R
\name{Intelligent_Investor}
\alias{Intelligent_Investor}
\title{Intelligent_Investor}
\usage{
Intelligent_Investor(Tickers)
}
\arguments{
\item{Tickers}{xxxxx}
}
\description{
Intelligent_Investor
}
