% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Investment_Horizon_Int.R
\name{Investment_Horizon_Int}
\alias{Investment_Horizon_Int}
\title{Investment_Horizon_Int}
\usage{
Investment_Horizon_Int(
  Tickers,
  RM,
  Rf,
  Initial_Date,
  Final_Date_Training,
  Final_Date,
  Frequency,
  Periodicity,
  Hidden,
  Stepmax,
  Asymmetry = "Negative",
  Type_ANNt = "T4",
  N_Assets,
  Base = "yahoo",
  Fun = "S_Out",
  Specific_Dates = Sys.Date(),
  Import = "Yes",
  Exclude_ticket = "",
  AQ = "A",
  Size = 2000,
  PE_Ratio = 15,
  PB_Ratio = 1.5,
  GI_min = 0,
  GI_max = 21.5,
  CR = 2,
  EPS = 0,
  Plot_IS = "Total Revenue",
  Plot_CF = "Cash Dividends Paid",
  Plot_BS = "Total Liabilities",
  Download = "Yes",
  Type_ANN = "ANNt",
  Order = "Yes",
  Continue_from = "1900-01-01"
)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets.}

\item{RM}{Proxy of the market.}

\item{Rf}{Risk free rate.}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day'). Assets with values not observed in the series are excluded.}

\item{Final_Date}{End date of the treatment series.}

\item{Frequency}{How many times simulate the investiment horizon.}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”.}

\item{Hidden}{Number of hidden neurons (If ” is the length series). For a good performance use '' to form a square input x hidden matrix of neurons.}

\item{Stepmax}{Number of replications per asset to train the ANN. For a good performance, use 7500.}

\item{Type_ANNt}{Select type ANNt: "T1"= NNet_Signal_Traning; "T2"= NNet_t_Training; "T3"= MC_Signal_Training; "T4"= MC_t_Training; "T5"= NNet_Signal_Test; "T6"= NNet_t_Test; "T7"= MC_Signal_Test; "T8"= Type_ANNt: MC_t_Test.}

\item{N_Assets}{Limit of asset numbers in the portfolio.}

\item{Base}{Database to use: "yahoo" or "Rus".}

\item{Fun}{Which technique to apply to generate portfolios:
'S_Out' uses the ANNt_Oliveira_Ceretta_S_Out function, this is the standard;
'Out' uses the ANNt_Oliveira_Cereta_Out function;
'S' uses the ANNt_Oliveira_Cereta_S function;
'Original' uses the ANNt_Oliveira_Ceretta function.}

\item{Import}{Import dates from external data base after first import in each re balance. "Yes"
or "No". "No" is the standard.}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}

\item{AQ}{Annual Statements (A) or Quarterly Statements (Q)}

\item{Size}{Minimum Total Assets in Millions}

\item{PE_Ratio}{Maximum Price/Earnings Ratio}

\item{PB_Ratio}{Maximum Price to Book Ratio}

\item{GI_min}{Graham indicator minimum}

\item{GI_max}{Graham indicator maximum}

\item{CR}{Minimum Current Ratio}

\item{EPS}{Earning per Share Minimum}

\item{Plot_IS}{Chart of Income Sheet account selected}

\item{Plot_CF}{Chart of Cash Flow account selected}

\item{Plot_BS}{Char of Balance Sheet account selected}

\item{Download}{Download asset prices for external data set or local data set. "Yes" or "No". "Yes" is the standard.}

\item{Type_ANN}{Select the network type: 'ANNt' or 'LSTMt' in RNN from ANNt}

\item{Order}{If "Yes" processes the asset selection, if "No" uses the already processed assets available in the database}

\item{Continue_from}{Determine if continue from a Specific_Date in the data}

\item{Initial_Date_Training}{Training series start Date.}

\item{Asymmetric}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative".}

\item{Specific_Date}{Specific dates for the end of training. Used to define
the investment horizon of portfolios from specific dates.}
}
\description{
Analyzes the return of portfolios with different investment horizons
}
\examples{
Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
Tickers <-c('AAPL','XOM','TSLA','KO', 'F')
RM <-c('^GSPC') #RM the S&P500
Rf <- 0
Initial_Date <-c('2024-01-03')
Final_Date_Training <-c('2024-06-03')
Final_Date <-c('2024-11-20')
Frequency <- 2
Periodicity <- c('daily')
Hidden <- 5
Stepmax <- 7500
Type_ANNt <- 'T8'
N_Assets <- 3
Base <- 'yahoo'
Fun <- 'S_Out'
Specific_Dates <- c(Sys.Date())
AQ='A'
Size=2000
PE_Ratio=150
PB_Ratio=11.5
GI_min=0
GI_max=2100.5
CR=1
EPS=0
Plot_IS=''
Plot_BS= ''
Plot_CF=''
Investment_Horizon_Int(Tickers <-c('AAPL','XOM','TSLA','KO', 'F'), RM <-c('^GSPC'),
Rf = 0,
Initial_Date =c('2024-01-03'),
Final_Date_Training =c('2024-06-03'),
Final_Date =c('2024-11-20'),
Frequency = 2,
Periodicity = c('daily'),
Hidden = 5,
Stepmax = 7500,
Asymmetry='Negative',
Type_ANNt = 'T8',
N_Assets = 3,
Base = 'yahoo',
Fun = 'S_Out',
Specific_Dates = c(Sys.Date()),
AQ='A',
Size=2000,
PE_Ratio=2000,
PB_Ratio=2000,
GI_min=-300,
GI_max=2100.5,
CR=0,
EPS=-300,
Plot_IS='',
Plot_BS= '',
Plot_CF='',
Download = 'Yes')
}
